refactor handleAPI function and write tests We have modularised our handleAPI function and refactored it, now it has a predictable return value and is testable. We have also written two tests, one for a valid input and one for an invalid input, the function also sanitises the input, hopefully preventing heroku from crashing relates to #47 #48
